+++ src/datasets/iterable_dataset.py
@@ -9,9 +9,9 @@ from typing import Any, Callable, Dict, Iterable, Iterator, List, Optional, Unio
 import numpy as np
 import pyarrow as pa
 
-from .arrow_dataset import DatasetInfoMixin, _align_features
-from .features import Features, Value
-from .features.features import FeatureType
+from .arrow_dataset import DatasetInfoMixin
+from .features import Features
+from .features.features import FeatureType, _align_features, _check_if_features_can_be_aligned
 from .formatting import PythonFormatter
 from .info import DatasetInfo
 from .splits import NamedSplit
@@ -1296,25 +1296,6 @@ def iterable_dataset(
     )
 
 
-def _check_if_features_can_be_aligned(features_list: List[Features]):
-    """Check if the dictionaries of features can be aligned.
-
-    Two dictonaries of features can be aligned if the keys they share have the same type or some of them is of type `Value("null")`.
-    """
-    name2feature = {}
-    for features in features_list:
-        for k, v in features.items():
-            if k not in name2feature or (isinstance(name2feature[k], Value) and name2feature[k].dtype == "null"):
-                name2feature[k] = v
-
-    for features in features_list:
-        for k, v in features.items():
-            if not (isinstance(v, Value) and v.dtype == "null") and name2feature[k] != v:
-                raise ValueError(
-                    f'The features can\'t be aligned because the key {k} of features {features} has unexpected type - {v} (expected either {name2feature[k]} or Value("null").'
-                )
-
-
 def _concatenate_iterable_datasets(
     dsets: List[IterableDataset],
     info: Optional[DatasetInfo] = None,
@@ -1354,7 +1335,7 @@ def _concatenate_iterable_datasets(
     # TODO: improve this to account for a mix of ClassLabel and Value for example
     # right now it would keep the type of the first dataset in the list
     features = Features(
-        {k: v for features in _align_features([dset.features for dset in dsets[::-1]]) for k, v in features.items()}
+        {k: v for features in _align_features([dset.features for dset in dsets]) for k, v in features.items()}
     )
 
     ex_iterables = [d._ex_iterable for d in dsets]
@@ -1407,7 +1388,7 @@ def _interleave_iterable_datasets(
     # TODO: improve this to account for a mix of ClassLabel and Value for example
     # right now it would keep the type of the first dataset in the list
     features = Features(
-        {k: v for features in _align_features([dset.features for dset in datasets[::-1]]) for k, v in features.items()}
+        {k: v for features in _align_features([dset.features for dset in datasets]) for k, v in features.items()}
     )
 
     ex_iterables = [d._ex_iterable for d in datasets]
